backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
selenium
webdriver-manager
pinecone-client
sentence-transformers
python-dotenv